refactor: /clients와 /group/.../clients 분리
